restructure modals + add these better modals to codelist form
