+++ app/components/codelist/confirm-delete-modal.hbs
@@ -1,32 +1,10 @@
-<AuModal
-  @title={{t 'utility.confirmation.body'}}
-  @modalOpen={{true}}
-  @closeModal={{fn @close false}} as |Modal|
->
-  <Modal.Body>
-    <p>
-      {{t
-        'codelist.crud.confirm-deletion'
-        name=@data.codelist.label
-        htmlSafe=true
-      }}
-    </p>
-  </Modal.Body>
-  <Modal.Footer>
-    <AuButton
-      @alert={{true}}
-      @loading={{@data.isLoading}}
-      @loadingText={{t 'utility.deleting'}}
-      {{on 'click' (fn @close true)}}
-    >
-      {{t 'codelist.crud.delete'}}
-    </AuButton>
-    <AuButton
-      @skin='secondary'
-      @disabled={{@data.isLoading}}
-      {{on 'click' (fn @close false)}}
-    >
-      {{t 'utility.cancel'}}
-    </AuButton>
-  </Modal.Footer>
-</AuModal>
+<Modals::BaseConfirmModal
+  @message={{
+    t 'codelist.crud.confirm-deletion' name=@data.codelist.label htmlSafe=true
+  }}
+  @alert={{true}}
+  @loading={{@data.isLoading}}
+  @loadingText={{t 'utility.deleting'}}
+  @confirmText={{t 'codelist.crud.delete'}}
+  @close={{@close}}
+/>
