+++ app/components/modals/confirm-discard-changes.hbs
@@ -1,0 +1,6 @@
+<Modals::BaseConfirm
+  @message={{t 'utility.confirm-leave-without-saving'}}
+  @alert={{true}}
+  @confirmMessage={{t 'utility.discard-changes'}}
+  @close={{@close}}
+/>
