+++ app/templates/edit.hbs
@@ -115,5 +115,5 @@
 </RdfaEditorContainer>
 <ConfirmRouteLeave
   @enabled={{this.dirty}}
-  @message={{t 'reglement-edit.confirm-leave-without-saving'}}
+  @message={{t 'utility.confirm-leave-without-saving'}}
 />
